+++ src/Zaya.ScreenTranslator.Impl.Shared/Logging/ILogConfigService.cs
@@ -1,0 +1,6 @@
+namespace Zaya.ScreenTranslator.Impl.Shared.Logging;
+
+public interface ILogConfigService
+{
+    LogConfig LoadOrCreate();
+}